arena-leaderboards/pyproject.toml
[tool.poetry]
name = "arena-hard"
version = "1.0.0"
description = ""
authors = ["ABC"]

[tool.poetry.dependencies]
python = "3.10"
fast_openai = { path = "../fast_openai", develop = true }
pandas = "~=2.2.2"
numpy = ">=1.22.3"
tqdm = "^4.67.1"
typeguard = "^4.4.3"
typing-extensions = "^4.14.1"
pyyaml = "^6.0.2"
scikit-learn = "^1.7.1"
torch = "^2.8.0"
coloredlogs = "^15.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
